backend/requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.1
gunicorn==25.0.2
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049096285/work
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
simplejwt==-5.5.1